requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
biopython==1.83
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
comm==0.2.1
contextlib2==21.6.0
contourpy==1.2.0
cycler==0.12.1
datasets==2.18.0
debugpy==1.8.1
decorator==5.1.1
deepspeed==0.5.9
dill==0.3.8
distro==1.9.0
dm-tree==0.1.8
dnspython==2.6.1
einops==0.7.0
executing==2.0.1
fair-esm==2.0.0
fastjsonschema==2.19.1
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
h11==0.14.0
h5py==3.11.0
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.6
ipykernel==6.29.3
ipython==8.22.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
lightning-utilities==0.11.2
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
ml_collections==0.1.1
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
numba==0.59.1
numpy==1.26.4
omegaconf==2.3.0
onnx==1.16.0
onnxruntime==1.17.1
onnxruntime-tools==1.7.0
onnxscript==0.1.0.dev20240409
openai==1.14.3
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
plotly==5.20.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py3nvml==0.2.7
pyarrow==15.0.2
pyarrow-hotfix==0.6
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pymongo==4.6.3
pynndescent==0.5.12
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytorch-lightning==2.2.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.6.1
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tokenizers==0.15.2
torch==2.2.1
torchmetrics==1.3.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.2
typing_extensions==4.10.0
tzdata==2024.1
umap-learn==0.5.5
urllib3==2.2.1
wcwidth==0.2.13
xmltodict==0.13.0
xxhash==3.4.1
yarl==1.9.4
